requirements.txt
playwright
requests
openpyxl
python-dotenv
colorama
